Fix occupation scaling (#8)
* fix incorrect DOM property label

I accidentally left it as "class=" when it should have been "className="

* Move header-related CSS into its own file

* fix occupation font size to be relative

when using rem, it seems the element wasn't resizing when resizing the window, so if I used a smaller window the occupation was huge, or a larger monitor the occupation was small. When I changed it to em which scales to the parent element it properly adjusted when I resized the window.